SumModel now implements CalculationModel interface to aid in dependency injection. New ProductModel class created with unit test that initially fails

+++ app/src/main/java/com/conradcrates/curvetechnicalchallenge/MainActivity.java
@@ -24,9 +24,9 @@ public class MainActivity extends AppCompatActivity implements Observer{
         GridLayout layout = (GridLayout)findViewById(R.id.layout_grid);
         result = (TextView)findViewById(R.id.text_sum);
 
-        List<Integer> values = new ArrayList<>();
+        List<Double> values = new ArrayList<>();
         for(int i = 0; i < 6; i++){
-            values.add(0);
+            values.add(0d);
         }
 
         viewModel = new MainActivityViewModel(values, new SumModel());
